the installer
All the functions/sections are functionally working.  the GetFileProductVersion function will be used in order to get the specific documentation for every Liberty Eiffel version (It is not used now due to the site's structure). 
Please note that the ProductVersion (row 19) will need to be changed accordingly for every version.